requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
open-clip-torch>=2.23.0
faiss-cpu>=1.7.4
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
gradio>=4.20.0
pyyaml>=6.0.1
tqdm>=4.66.0
pillow>=10.2.0
openai>=1.12.0
datasets>=2.17.0
